test(cli): consolidate markdown coverage and docs
